Screen size letterboxing works. Added pretty stuff to homescreen Fixed backstack

+++ core/src/com/rockets/gamescreen/world/PhysicalEntity.java
@@ -115,7 +115,7 @@ public abstract class PhysicalEntity extends GameEntity implements PhysicsEntity
     protected void setBounciness(float bounciness){
         this.bounciness = bounciness;
     }
-    protected void setFriction(int friction) {
+    protected void setFriction(float friction) {
         this.friction = friction;
     }
